# docs/30-api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dindinho API",
    "description": "API para o app de finanças pessoais Dindinho.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/accounts/{id}": {
      "patch": {
        "summary": "Atualizar conta",
        "tags": ["accounts"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "color": {
                    "type": "string",
                    "minLength": 4
                  },
                  "icon": {
                    "type": "string",
                    "minLength": 1
                  },
                  "closingDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  },
                  "dueDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  },
                  "limit": {
                    "type": "number",
                    "exclusiveMinimum": true,
                    "minimum": 0,
                    "nullable": true
                  },
                  "brand": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "color": {
                      "type": "string"
                    },
                    "icon": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string",
                      "enum": ["STANDARD", "CREDIT"]
                    },
                    "ownerId": {
                      "type": "string"
                    },
                    "balance": {
                      "type": "number",
                      "default": 0
                    },
                    "creditCardInfo": {
                      "type": "object",
                      "properties": {
                        "closingDay": {
                          "type": "number"
                        },
                        "dueDay": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number",
                          "nullable": true
                        },
                        "availableLimit": {
                          "type": "number",
                          "nullable": true
                        },
                        "brand": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": ["closingDay", "dueDay"],
                      "additionalProperties": false,
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "color",
                    "icon",
                    "type",
                    "ownerId",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "pattern": "^[A-Z0-9_]+$"
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "array",
                            "items": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                }
                              ]
                            }
                          }
                        },
                        "required": ["code", "message", "path"],
                        "additionalProperties": true
                      }
                    },
                    "details": {}
                  },
                  "required": ["statusCode", "error", "message"],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "pattern": "^[A-Z0-9_]+$"
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "array",
                            "items": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                }
                              ]
                            }
                          }
                        },
                        "required": ["code", "message", "path"],
                        "additionalProperties": true
                      }
                    },
                    "details": {}
                  },
                  "required": ["statusCode", "error", "message"],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "pattern": "^[A-Z0-9_]+$"
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "array",
                            "items": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                }
                              ]
                            }
                          }
                        },
                        "required": ["code", "message", "path"],
                        "additionalProperties": true
                      }
                    },
                    "details": {}
                  },
                  "required": ["statusCode", "error", "message"],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "pattern": "^[A-Z0-9_]+$"
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "array",
                            "items": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                }
                              ]
                            }
                          }
                        },
                        "required": ["code", "message", "path"],
                        "additionalProperties": true
                      }
                    },
                    "details": {}
                  },
                  "required": ["statusCode", "error", "message"],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "pattern": "^[A-Z0-9_]+$"
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "array",
                            "items": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                }
                              ]
                            }
                          }
                        },
                        "required": ["code", "message", "path"],
                        "additionalProperties": true
                      }
                    },
                    "details": {}
                  },
                  "required": ["statusCode", "error", "message"],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "pattern": "^[A-Z0-9_]+$"
                    },
                    "requestId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "array",
                            "items": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "number"
                                }
                              ]
                            }
                          }
                        },
                        "required": ["code", "message", "path"],
                        "additionalProperties": true
                      }
                    },
                    "details": {}
                  },
                  "required": ["statusCode", "error", "message"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}